Warn if Ecto schemas aren't kept with only other schemas
This is a preliminary implementation of this feature. The determination
of whether a file defines an Ecto schema or not is pretty naive, but it
might be enough. We'll see if we run into edge cases when it's added to
a real application and then we'll update as we get more information.