Add check for documenting GraphQL schemas
This is a first take a a check to require documentation for graphql
schemas. Right now it assumes that the following functions are only used
for defining parts of a GraphQL schema:

```
[:field, :scalar, :arg, :object, :input_object, :directive, :interface,
:union, :enum, :value]
```

If we end up with functions that have the same names then we can ingore
just the calls to those specific functions, or we can make this check a
bit more robust, but it would be good to get the context for the
real-world failure before making those changes.