Fix bug with GraphQL schema documentation check
Well, the assumption that `field` would always be used in Absinthe type
definitions was a pretty short lived idea! I totally forgot that Ecto
schemas also had a `field` macro, and so we pretty quickly got false
positives in here. This now fixes that to only warn if the `field` macro
is used inside one of the other "parent" macros like `object`,
`input_object` or `interface`.