Update to `credo ~> 1.6` and fix compilation warnings
This commit updates the version of `Credo` that we depend on to `~> 1.6`
and fixes any compilation warnings that result from the upgrade.

This commit also deprecates the use of `Nicene` as a plugin. At the
moment, when used as a plugin Nicene will opt the user in to several
checks with default configuration. The user should instead opt-in to
checks by listing them explicitly in their `.credo.exs`. `Nicene.init/1`
is maintained for now to ensure that folks who are using `Nicene` as a
plugin will get a deprecation warning rather than breaking their builds.

It also fixes issues highlighted by `mix format` and `mix credo`.